aide-data-engine/pyproject.toml
[tool.poetry]
name = "aide-data-engine"
version = "0.1.0"
description = "Data processing pipeline for AIDE Platform - embedding, deduplication, AI classification"
authors = ["AIDE Team <team@aide-platform.com>"]
readme = "README.md"
packages = [{include = "aide_data_engine"}]

[tool.poetry.dependencies]
python = "^3.11"
aide-data-core = {path = "../aide-data-core", develop = true}
sentence-transformers = "^2.2.2"
scikit-learn = "^1.3.0"
anthropic = "^0.18.0"
numpy = "^1.24.0"
pandas = "^2.0.0"
pydantic = "^2.0.0"
sqlalchemy = "^2.0.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.4.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=aide_data_engine"
